+++ example/example/urls.py
@@ -4,6 +4,7 @@ from django.conf import settings
 from django.contrib import admin
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
 from django.conf.urls.static import static
+from django.views.generic import TemplateView
 
 admin.autodiscover()
 
@@ -11,6 +12,7 @@ urlpatterns = patterns('',
     url(r'^ska/', include('ska.contrib.django.ska.urls')),
     url(r'^foo/', include('foo.urls')),
     url(r'^admin/', include(admin.site.urls)),
+    url(r'^$', TemplateView.as_view(template_name='home.html')),
 )
 
 if settings.DEBUG: