+++ install.sh
@@ -1,3 +1,3 @@
 reset
 python setup.py install
-mkdir -p example/db example/static example/tmp example/media example/media/foo-images example/media/static
+mkdir -p example/logs example/db example/static example/tmp example/media example/media/foo-images example/media/static