+++ src/ska/defaults.py
@@ -9,9 +9,13 @@ Application defaults.
 - `DEFAULT_VALID_UNTIL_PARAM` (str): Default name of the GET param holding the ``valid_until`` value.
   Default value is `valid_until`.
 - `DEFAULT_EXTRA_PARAM` (str): Default name of the GET param holding the ``extra`` value.
-  Default value is `extra_keys`.
+  Default value is `extra`.
+- `DEFAULT_PROVIDER_PARAM` (str): Default name of the GET param holding the ``provider`` value.
+  Default value is `provider`.
 - `DEFAULT_URL_SUFFIX` (str): Suffix to add after the ``endpoint_url`` and before the appended signature
   params.
+- `DEFAULT_RESERVED_PARAMS` (list): List of GET params reserved by default. Users should not be allowed
+  to use them.
 """
 
 __title__ = 'ska.defaults'
@@ -20,7 +24,8 @@ __copyright__ = 'Copyright (c) 2013 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'TIMESTAMP_FORMAT', 'SIGNATURE_LIFETIME', 'DEFAULT_SIGNATURE_PARAM', 'DEFAULT_AUTH_USER_PARAM',
-    'DEFAULT_VALID_UNTIL_PARAM', 'DEFAULT_EXTRA_PARAM', 'DEFAULT_URL_SUFFIX', 'DEBUG'
+    'DEFAULT_VALID_UNTIL_PARAM', 'DEFAULT_EXTRA_PARAM', 'DEFAULT_PROVIDER_PARAM', 'DEFAULT_URL_SUFFIX',
+    'DEFAULT_RESERVED_PARAMS', 'DEBUG'
     )
 
 TIMESTAMP_FORMAT = "%Y-%m-%dT%H:%M:%S"
@@ -37,10 +42,22 @@ DEFAULT_AUTH_USER_PARAM = 'auth_user'
 # Default name of the GET param holding the ``valid_until`` value.
 DEFAULT_VALID_UNTIL_PARAM = 'valid_until'
 
-# Default name of the GET param holding the ``extra_keys`` value.
+# Default name of the GET param holding the ``extra`` value.
 DEFAULT_EXTRA_PARAM = 'extra'
+
+# Default name of the GET param holding the ``provider`` value.
+DEFAULT_PROVIDER_PARAM = 'provider'
 
 # Suffix to add after the ``endpoint_url`` and before the appended signature params.
 DEFAULT_URL_SUFFIX = '?'
+
+# A list of GET params reserved by default. Users should not be allowed to use them.
+DEFAULT_RESERVED_PARAMS = (
+    DEFAULT_SIGNATURE_PARAM,
+    DEFAULT_AUTH_USER_PARAM,
+    DEFAULT_VALID_UNTIL_PARAM,
+    DEFAULT_EXTRA_PARAM,
+    DEFAULT_PROVIDER_PARAM
+)
 
 DEBUG = False